# Conflicts:
#	src/app/map-editor/tool-bar/tools/device-placer/device-placer.component.ts
#	src/app/map-editor/tool-bar/tools/device-placer/list/device-placer.list.ts
#	src/assets/i18n/en.json

+++ src/app/map-editor/tool-bar/tools/device-placer/device-placer.component.ts
@@ -29,20 +29,25 @@ import {ModelsConfig} from '../../../../map/models/models.config';
   templateUrl: './device-placer.html'
 })
 export class DevicePlacerComponent implements Tool, OnInit, OnDestroy {
+  private static defaultHeight = 100;
+
   active: boolean = false;
   disabled: boolean = true;
+  activeDevice: SinkBag | AnchorBag;
+  displayHeightDialog: boolean = false;
+  heightInCentimeters: number = DevicePlacerComponent.defaultHeight;
   private subscriptionDestroyer: Subject<void> = new Subject<void>();
   private map: d3.selection;
   private scale: Scale;
   private floorId: number;
   private scaleCalculations: ScaleCalculations;
   private sinks: SinkBag[] = [];
-  private activeDevice: SinkBag | AnchorBag;
   private draggedDevice: DeviceDto;
   private contextMenu: DeviceCallbacks;
   private confirmationBody: string;
   private containerBox: Box;
   private properPlacementErrorMessage: string;
+  private devicesToAdd: [SinkBag, AnchorBag];
   private devicesOnMap: Anchor[] = [];
 
   private static average(data: number[]): number {
@@ -68,8 +73,12 @@ export class DevicePlacerComponent implements Tool, OnInit, OnDestroy {
 
   ngOnInit() {
     this.contextMenu = {
-      unset: () => {
+      remove: () => {
         this.removeFromMap();
+      },
+      edit: () => {
+        this.displayHeightDialog = true;
+        this.heightInCentimeters = this.activeDevice.deviceInList.z;
       }
     };
     this.bindMapSelection();
@@ -127,6 +136,40 @@ export class DevicePlacerComponent implements Tool, OnInit, OnDestroy {
       this.map = mapLoaded.container;
       this.containerBox = this.map.node().getBBox();
     });
+  }
+
+
+  heightDialogClosed(confirmed: boolean, isValid: boolean): void {
+    if (!isValid) {
+      return;
+    }
+    if (!confirmed && !!this.devicesToAdd) {
+      this.removeFromMap(false);
+    } else if (!!this.devicesToAdd) {
+      const [sinkToAdd, anchorToAdd] = this.devicesToAdd;
+      if (this.draggedDevice.type === DeviceType.SINK) {
+        sinkToAdd.deviceInEditor.activateForMouseEvents();
+        sinkToAdd.deviceInEditor.contextMenuOn(this.contextMenu);
+        sinkToAdd.deviceInList.z = this.heightInCentimeters;
+        this.configurationService.addSink(<Sink>this.updateDevicePosition(sinkToAdd));
+      } else if (this.draggedDevice.type === DeviceType.ANCHOR) {
+        anchorToAdd.deviceInEditor.activateForMouseEvents();
+        anchorToAdd.deviceInEditor.contextMenuOn(this.contextMenu);
+        anchorToAdd.deviceInList.z = this.heightInCentimeters;
+        this.configurationService.addAnchor(<Sink>sinkToAdd.deviceInList, <Anchor>this.updateDevicePosition(anchorToAdd));
+      }
+    } else if (confirmed) {
+      this.activeDevice.deviceInList.z = this.heightInCentimeters;
+      if (this.activeDevice.deviceInEditor.type === DeviceType.ANCHOR) {
+        this.configurationService.updateAnchor(this.activeDevice.deviceInList);
+      } else if (this.activeDevice.deviceInEditor.type === DeviceType.SINK) {
+        this.configurationService.updateSink(<Sink>this.activeDevice.deviceInList);
+      }
+    }
+
+    this.activeDevice.deviceInEditor.updateHeight(this.activeDevice.deviceInList.z);
+    this.displayHeightDialog = false;
+    this.devicesToAdd = null;
   }
 
   private captureScaleChanges(): void {
@@ -194,13 +237,13 @@ export class DevicePlacerComponent implements Tool, OnInit, OnDestroy {
     this.devicePlacerService.onDroppedInside.takeUntil(this.subscriptionDestroyer).subscribe((coordinates: Point): void => {
       const dropTransitionCoordinates = this.zoomService.calculateTransition(coordinates);
       if (!!this.draggedDevice) {
+        this.heightInCentimeters = this.draggedDevice.device.z == null ? DevicePlacerComponent.defaultHeight : this.draggedDevice.device.z;
+        this.displayHeightDialog = true;
         this.devicesOnMap.push(this.draggedDevice.device);
         if (this.draggedDevice.type === DeviceType.SINK) {
           const sinkBag: SinkBag = this.placeSinkOnMap(<Sink>this.draggedDevice.device, dropTransitionCoordinates);
-          sinkBag.deviceInEditor.activateForMouseEvents();
-          sinkBag.deviceInEditor.contextMenuOn(this.contextMenu);
           this.devicePlacerService.emitActivated(sinkBag.deviceInEditor);
-          this.configurationService.addSink(<Sink>this.updateDevicePosition(sinkBag));
+          this.devicesToAdd = [sinkBag, null];
         } else if (this.draggedDevice.type === DeviceType.ANCHOR) {
           if (this.activeDevice.deviceInEditor.type === DeviceType.ANCHOR) {
             const index: number = this.sinks.findIndex((sink: SinkBag): boolean => {
@@ -210,10 +253,8 @@ export class DevicePlacerComponent implements Tool, OnInit, OnDestroy {
           }
           const sinkBag: SinkBag = <SinkBag>this.activeDevice;
           const anchorBag: AnchorBag = this.placeAnchorOnMap(sinkBag, <Anchor>this.draggedDevice.device, dropTransitionCoordinates);
-          anchorBag.deviceInEditor.activateForMouseEvents();
-          anchorBag.deviceInEditor.contextMenuOn(this.contextMenu);
           this.devicePlacerService.emitActivated(anchorBag.deviceInEditor);
-          this.configurationService.addAnchor(<Sink>sinkBag.deviceInList, <Anchor>this.updateDevicePosition(anchorBag));
+          this.devicesToAdd = [sinkBag, anchorBag];
         }
 
         this.checkProperPlacement();
@@ -289,7 +330,7 @@ export class DevicePlacerComponent implements Tool, OnInit, OnDestroy {
     const sinkDrawConfiguration: DeviceInEditorConfiguration = {
       id: `sink-${sink.shortId}`,
       clazz: `sink`,
-      heightInMeters: sink.z
+      height: sink.z
     };
     const sinkOnMap: SinkInEditor = new SinkInEditor(
       sink.shortId,
@@ -316,7 +357,7 @@ export class DevicePlacerComponent implements Tool, OnInit, OnDestroy {
     const anchorDrawConfiguration: DeviceInEditorConfiguration = {
       id: `anchor-${anchor.shortId}`,
       clazz: `anchor`,
-      heightInMeters: anchor.z
+      height: anchor.z
     };
     const anchorInEditor: AnchorInEditor = new AnchorInEditor(
       anchor.shortId,
@@ -339,7 +380,7 @@ export class DevicePlacerComponent implements Tool, OnInit, OnDestroy {
     return anchorBag;
   }
 
-  private removeFromMap(): void {
+  private removeFromMap(removeFromConfiguration: boolean = true): void {
     this.devicesOnMap = this.devicesOnMap.filter((anchor: Anchor) => {
       return anchor.shortId !== this.activeDevice.deviceInList.shortId;
     });
@@ -368,7 +409,9 @@ export class DevicePlacerComponent implements Tool, OnInit, OnDestroy {
         return sink.deviceInEditor.hasAnchor(anchorBag);
       });
       this.removeAnchorFromSink(sinkWithAnchor, anchorBag);
-      this.configurationService.removeAnchor(anchorBag.deviceInList);
+      if (removeFromConfiguration) {
+        this.configurationService.removeAnchor(anchorBag.deviceInList);
+      }
     }
     this.devicePlacerService.emitMapModeActivated();
     this.sinks.forEach((sinkBag: SinkBag) => {
